make a legacy folder
