--- src/app/api/models/rbacModeNotReadyError.ts
@@ -1,11 +1,0 @@
-/***********************************************************
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License
- **********************************************************/
-export class RbacModeNotReadyError extends Error {
-    constructor(message?: string) {
-        super(message);
-        Object.setPrototypeOf(this, new.target.prototype);
-        this.name = RbacModeNotReadyError.name;
-    }
-}
